note_encryption: extract Domain impl, OrchardDoimain and NoteBytes definitions into separate modules in note_encryption subfolder
